+++ scripts/systems/dnd5e.js
@@ -6,8 +6,16 @@ export default {
     "ITEM_QUANTITY_ATTRIBUTE": "data.quantity",
 
     // Item types and the filters actively remove items from the item pile inventory UI that users cannot loot, such as spells, feats, and classes
-    "ITEM_TYPE_ATTRIBUTE": "type",
-    "ITEM_TYPE_FILTERS": "spell,feat,class",
+    "ITEM_FILTERS": [
+        {
+            "path": "type",
+            "filters": "spell,feat,class"
+        },
+        {
+            "path": "data.weaponType",
+            "filters": "natural"
+        }
+    ],
 
     // Dynamic attributes are things like currencies or transferable powers that exist as editable number fields on character sheets
     "DYNAMIC_ATTRIBUTES": [